refactor html structure scanner
